server/requirements.txt
bcrypt==2.0.0
Cerberus==0.9.2
cffi==1.6.0
Django==1.9.7
djangorestframework==3.3.3
drf-nested-routers==0.11.1
pyaml==15.8.2
pycparser==2.14
pystalkd==1.2.3
PyYAML==3.11
six==1.10.0
